Set up screen and line from floor to center in diagram-display.
New property in the configuration-context to hold the minimum distance from the screen to the floor.

Initialize the minimum distance to floor in equipment-form.

Within configuration-form, changing the screen or orientation will re-initialize the distance and minimum distance from floor values.

When the user attempts to change the distance from floor to screen center in the config form, they cannot change the value below the minimum distance.